Inserted background mountains and path on front page
Inserted svg files with mountains and path in html. Put id's on the divs and the img. Styled in css with relative positions on the div elements and absolute positions on the img files. Created a mid point in the div with "left: 50%". The img files styled and placed according to the mid point in the div element.
Added z-index to the img files, so they lay in the correct order (behind the main elements, as a background).