Created hobby section in about.html
Created hobby section with headline and hobby box including image with iluustrations positioned with absolute, a sub-headline and a paragraph. Struggled with positioning the arrows and dots because of absolute position because padding and margin wouldn't work. Decided to place it with use of position of first and last child.